tested new logic in advertisement service and some refactoring

+++ TradeZoneBackend/src/main/java/TradeZone/data/model/rest/ProfileUpdate.java
@@ -13,7 +13,7 @@ public class ProfileUpdate {
 
     private String lastName;
 
-    private int town;
+    private Long town;
 
     private String aboutMe;
 }
